+++ python-package/HISTORY.md
@@ -9,6 +9,8 @@ Unreleased
   * `nichtparasoup.core.imagecrawler.RemoteFetcher` got the ability to write communication logs.  
     The log target dir is controlled via env var `NP_DEBUG_REMOTEFETCHER_STOREDIR`.  
     The log target dir must exist and be writeable.
+  * Crawler `nichtparasoup.imagecrawlers.instagram.InstagramProfile` got a new optional config: `profile_id`.  
+    See the [InstagramProfile's docs](docs/imagecrawlers/instagram_profile.md) for details.
 * Misc
   * Tests can persist logs of `nichtparasoup` - controlled via env var `NP_TESTLOG_NAME`.  
     If the env var is present: